Fixed the left-right issue
So, this Newton guy had something to say about for every action there's
an opposite action. If the watch or phone hits 10G in the direction
opposite of activation it is going to hit 10G when it comes to rest. So,
the code ignores inputs in the wrong direction as long as it does one in
the proper direction.

Seems to solve the problem!